pyproject.toml
[project]
name = "thehackerlibrary"
version = "0.1.0"
description = "The Hacker Library Post Scraper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "newspaper3k>=0.2.8",
    "lxml[html-clean]>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "sqlalchemy>=2.0.43",
    "sqlalchemy-utils>=0.42.0",
    "tabulate>=0.9.0",
    "aiohttp>=3.13.0",
]

[project.scripts]
thehackerlibrary = "thehackerlibrary.__main__:main"

[dependency-groups]
dev = [
    "sqlacodegen>=3.0.0",
]
